Notebook for suborbital flight velocity optimization
This has a sufficiently finalized answer for my purposes.
Under consideration is the optimal velocity of a suborbital
flight, assuming they have the ability to exert a continuous
upwards thrust through the stable part of the trajectory.

The distance range under which this differs from the full
orbital velocity is the main product of this notebook.